servers/gemini-audio/pyproject.toml
[project]
name = "gemini-audio-mcp-server"
version = "1.0.0"
description = "Audio Understanding MCP Server using Google Gemini API"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "google-generativeai>=0.8.0",
    "mcp[cli]>=1.9.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
